Rebuild the docs
It looks like we'd mistakenly made some direct changes to the HTML, rather than updating the Markdown source and then rebuilding the HTML with `npm run make`.

This commit updates the Markdown with changes to the HTML so that they
are both aligned, and everything is up-to-date with `npm run make`.